fix: consistently pass blocks from leaf templates upwards
The passing of blocks through more than one template to its base was
broken because templates between the leaf and the root were not
keeping blocks from their leaves.

This change fixes that issue by copying blocks received recursively
into the set of blocks collected at the current template and then
making sure we're using the blocks received recursively at the root
template.

The function templatel--compiler-wrap was renamed to
templatel--compiler-code because it seemed to more accurately describe
what it generates.