refactoring: cleaner template creation/import
We didn't ever use the wrapping assoc list used on the return of
`templatel-new`, so that was removed;

The wrapping of the compiled code with the runtime function
`templatel--compiler-code' has been moved off the template creation
too because than one compiled block of code can be wrapped by
different runtime functions.  This might help implementing the block
rendering traversal.  If it doesn't it might be changed back to how it
looked before this change.