fix: super() inject rendered parent's block with the same name
This took quite a bit of refactoring but it seems to work.  The most
important part of the change was the split of runtimes for blocks and
templates.

With such a split, the parent's blocks are rendered from the root
template to the leaf before any rendering of any templates; once all
the blocks are collected, the rendering of the leaf template starts
carring all the parent blocks found by the process triggered by
extends.

This design might have to evolve and more corners might have to be
polished; although the general idea seems to work, we still need more
test cases and more error handling to have a good experience while
editing so one doesn't waste time with awful error messages on typos
or such.