fix: Make sure compiling a block and rendering are separate actions
The code for compiling a block lives very close to the code that
renders the block and puts it into the `rt/blocks' map.  And the
stitching of teh `body' parameter into the compiling of the block was
spilling over the code that renders it, which was fixed by surrounding
the a unquoting of the variable with `progn', so even if the code
changes in strange ways, the spilling over won't happen again.