feat/refactor: tests via `is` operator & undefined value
This is somewhat a big change.  The most noticeable change is that
variables or attributes can be `undefined'.  This new type of value is
being introduced so template rendering doesn't fail and it gives users
the chance to test if variables are defined or not.  E.g.:

  (templatel-render-string "{{ something }}" '())

The above snippet will render the string `undefined'.  If one wants to
test if the variable is defined, they can use the following code:

  {% if something is defined %}
     {{ something }}
  {% endif %}

Notice that although `{% if something %}` will evaluate to `nil' if
the variable is not defined, it would also evaluate to true if the
variable was defined but had its contents set to nil.

Notice that the same behavior now applies to accessing attributes as
well. e.g.:

  (templatel-render-string "{{ user.city }}"
                           '(("user" . (("name" . "GNU")))))

The above snippet will also render the string `undefined', since the
sub-list doesn't contain the entry `city'.  Both the `defined` test
and the `if` statement behave the same way for attributes.

Alongside with the changes in the language, the API also got two new
functions `templatel-env-add-test' and `templatel-env-remove-test'
allowing the use of custom defined tests.