fix: for loop must access expression via value stack
The compilation of for loop statements was accessing the expression
that evaluates what collection it will iterate over via a recursive
call to the compiler, and that only worked for other cases by luck.

In order to test this feature, I've added a new filter: `sort`, it
doesn't take any arguments for now but could be a bit more flexible.

Refs: #11